+++ declaration.go
@@ -10,17 +10,34 @@ import (
 	"sync"
 )
 
-// declaration is an interface that describes how to build a dependency.
+// declaration is an interface that describes how to build a value of a specific
+// type.
 type declaration interface {
-	getType() reflect.Type
-	location() (string, int)
+	// GetType returns the type of the value constructed by this declaration.
+	GetType() reflect.Type
+
+	// Location returns the location of the declaration in code.
+	Location() (string, int)
+
+	// IsDependency returns true if other declarations depend upon this one.
+	IsDependency() bool
+
+	// Dependencies returns the declarations that this declaration depends upon,
+	// sorted by type.
+	Dependencies() []declaration
+
+	// dependsOn returns true if d depends on t, whether directly or indirectly.
 	dependsOn(t declaration, cycle []declaration) ([]declaration, bool)
+
+	// markAsDependency marks the declaration as a dependency. That is, other
+	// declarations depend upon this one.
+	markAsDependency()
 }
 
 // constructor is a function that constructs values of type T.
 type constructor[T any] func(*Context) (T, error)
 
-// declarationOf describes how to build dependencies of a specific type.
+// declarationOf describes how to build values of type T.
 type declarationOf[T any] struct {
 	m sync.Mutex
 
@@ -28,11 +45,13 @@ type declarationOf[T any] struct {
 	line int
 
 	isConstructed bool
-	dependencies  map[reflect.Type]declaration
+	deps          map[reflect.Type]declaration
+	isDep         bool
 	construct     constructor[T]
 	value         T
 }
 
+// Declare declares a constructor for values of type T.
 func (d *declarationOf[T]) Declare(
 	decl func() (constructor[T], error),
 ) error {
@@ -55,13 +74,14 @@ func (d *declarationOf[T]) Declare(
 	return nil
 }
 
+// AddDependency marks t as a dependency of d.
 func (d *declarationOf[T]) AddDependency(t declaration) error {
-	if t == d {
-		file, line := d.location()
+	if t.GetType() == d.GetType() {
+		file, line := d.Location()
 
 		return fmt.Errorf(
 			"constructor for %s (%s:%d) depends on itself",
-			d.getType(),
+			d.GetType(),
 			filepath.Base(file),
 			line,
 		)
@@ -70,16 +90,16 @@ func (d *declarationOf[T]) AddDependency(t declaration) error {
 	if cycle, ok := t.dependsOn(d, nil); ok {
 		message := fmt.Sprintf(
 			"constructor for %s introduces a cyclic dependency:",
-			d.getType(),
+			d.GetType(),
 		)
 
 		for i := len(cycle) - 1; i >= 0; i-- {
 			dep := cycle[i]
-			file, line := dep.location()
+			file, line := dep.Location()
 
 			message += fmt.Sprintf(
 				"\n\t-> %s (%s:%d)",
-				dep.getType(),
+				dep.GetType(),
 				filepath.Base(file),
 				line,
 			)
@@ -91,15 +111,20 @@ func (d *declarationOf[T]) AddDependency(t declaration) error {
 	d.m.Lock()
 	defer d.m.Unlock()
 
-	if d.dependencies == nil {
-		d.dependencies = map[reflect.Type]declaration{}
+	if d.deps == nil {
+		d.deps = map[reflect.Type]declaration{}
 	}
 
-	d.dependencies[t.getType()] = t
+	d.deps[t.GetType()] = t
+	t.markAsDependency()
 
 	return nil
 }
 
+// Resolve returns the value constructed by this declaration.
+//
+// The constructor is called only once. Subsequent calls to Resolve() return the
+// same value.
 func (d *declarationOf[T]) Resolve(ctx *Context) (T, error) {
 	d.m.Lock()
 	defer d.m.Unlock()
@@ -111,7 +136,7 @@ func (d *declarationOf[T]) Resolve(ctx *Context) (T, error) {
 	if d.construct == nil {
 		panic(fmt.Sprintf(
 			"no constructor is declared for %s",
-			d.getType(),
+			d.GetType(),
 		))
 	}
 
@@ -127,39 +152,67 @@ func (d *declarationOf[T]) Resolve(ctx *Context) (T, error) {
 	}
 
 	d.isConstructed = true
-	d.dependencies = nil
 	d.construct = nil
 	d.value = v
 
 	return v, nil
 }
 
-func (d *declarationOf[T]) getType() reflect.Type {
+// GetType returns the type of the value constructed by this declaration.
+func (d *declarationOf[T]) GetType() reflect.Type {
 	return typeOf[T]()
 }
 
-func (d *declarationOf[T]) location() (string, int) {
+// Location returns the location of the declaration in code.
+func (d *declarationOf[T]) Location() (string, int) {
 	d.m.Lock()
 	defer d.m.Unlock()
 
 	return d.file, d.line
 }
 
+// IsDependency returns true if other declarations depend upon this one.
+func (d *declarationOf[T]) IsDependency() bool {
+	d.m.Lock()
+	defer d.m.Unlock()
+
+	return d.isDep
+}
+
+// Dependencies returns the declarations that this declaration depends upon,
+// sorted by type.
+func (d *declarationOf[T]) Dependencies() []declaration {
+	d.m.Lock()
+	defer d.m.Unlock()
+
+	return sortDeclarations(d.deps)
+}
+
+// dependsOn returns true if d depends on t, whether directly or indirectly.
 func (d *declarationOf[T]) dependsOn(t declaration, cycle []declaration) ([]declaration, bool) {
-	if t.getType() == d.getType() {
+	if t.GetType() == d.GetType() {
 		return append(cycle, d), true
 	}
 
 	d.m.Lock()
 	defer d.m.Unlock()
 
-	for _, dep := range d.dependencies {
+	for _, dep := range d.deps {
 		if cycle, ok := dep.dependsOn(t, cycle); ok {
 			return append(cycle, d), true
 		}
 	}
 
 	return nil, false
+}
+
+// markAsDependency marks the declaration as a dependency. That is, other
+// declarations depend upon this one.
+func (d *declarationOf[T]) markAsDependency() {
+	d.m.Lock()
+	defer d.m.Unlock()
+
+	d.isDep = true
 }
 
 // findLocation returns the file and line number of the first frame in the